requirements.txt
flask
openai
PyPDF2
python-docx
gunicorn
os
python-dotenv